Break out TabHolder from Common/UI/ViewGroup.cpp. A few new atlas images
