+++ Common/UI/ViewGroup.h
@@ -260,84 +260,6 @@ public:
 	std::string DescribeText() const override;
 };
 
-class ChoiceStrip : public LinearLayout {
-public:
-	ChoiceStrip(Orientation orientation, LayoutParams *layoutParams = 0);
-
-	void AddChoice(std::string_view title);
-	void AddChoice(ImageID buttonImage);
-
-	int GetSelection() const { return selected_; }
-	void SetSelection(int sel, bool triggerClick);
-
-	void EnableChoice(int choice, bool enabled);
-
-	bool Key(const KeyInput &input) override;
-
-	void SetTopTabs(bool tabs) { topTabs_ = tabs; }
-
-	std::string DescribeLog() const override { return "ChoiceStrip: " + View::DescribeLog(); }
-	std::string DescribeText() const override;
-
-	Event OnChoice;
-
-private:
-	StickyChoice *Choice(int index);
-	void OnChoiceClick(EventParams &e);
-
-	int selected_ = 0;   // Can be controlled with L/R.
-	bool topTabs_ = false;
-};
-
-class TabHolder : public LinearLayout {
-public:
-	TabHolder(Orientation orientation, float stripSize, View *bannerView, LayoutParams *layoutParams = 0);
-
-	template <class T>
-	T *AddTab(std::string_view title, T *tabContents) {
-		AddTabContents(title, tabContents);
-		return tabContents;
-	}
-	void AddTabDeferred(std::string_view title, std::function<ViewGroup *()> createCb);
-	void EnableTab(int tab, bool enabled) {
-		tabStrip_->EnableChoice(tab, enabled);
-	}
-
-	void AddBack(UIScreen *parent);
-
-	// Returns true if the tab wasn't created before (but is now).
-	bool SetCurrentTab(int tab, bool skipTween = false);
-
-	int GetCurrentTab() const { return currentTab_; }
-	std::string DescribeLog() const override { return "TabHolder: " + View::DescribeLog(); }
-
-	void PersistData(PersistStatus status, std::string anonId, PersistMap &storage) override;
-
-	void EnsureAllCreated();
-
-	LinearLayout *Container() { return tabContainer_; }
-
-	const std::vector<ViewGroup *> &GetTabContentViews() const {
-		return tabs_;
-	}
-
-private:
-	void AddTabContents(std::string_view title, ViewGroup *tabContents);
-	void OnTabClick(EventParams &e);
-	bool EnsureTab(int index);  // return true if it actually created a tab.
-
-	View *bannerView_ = nullptr;
-	LinearLayout *tabContainer_ = nullptr;
-	ChoiceStrip *tabStrip_ = nullptr;
-	ScrollView *tabScroll_ = nullptr;
-	ViewGroup *contents_ = nullptr;
-
-	int currentTab_ = 0;
-	std::vector<ViewGroup *> tabs_;
-	std::vector<AnchorTranslateTween *> tabTweens_;
-	std::vector<std::function<ViewGroup *()>> createFuncs_;
-};
-
 class CollapsibleHeader;
 
 class CollapsibleSection : public LinearLayout {
