+++ UI/DisplayLayoutScreen.cpp
@@ -24,8 +24,8 @@
 #include "Common/UI/Context.h"
 #include "Common/UI/View.h"
 #include "Common/UI/UIScreen.h"
+#include "Common/UI/TabHolder.h"
 #include "Common/Math/math_util.h"
-#include "Common/System/Display.h"
 #include "Common/System/NativeApp.h"
 #include "Common/VR/PPSSPPVR.h"
 #include "Common/StringUtils.h"
