+++ UI/MainScreen.h
@@ -23,6 +23,7 @@
 #include "Common/File/Path.h"
 #include "Common/UI/UIScreen.h"
 #include "Common/UI/ViewGroup.h"
+#include "Common/UI/TabHolder.h"
 #include "UI/MiscScreens.h"
 #include "Common/File/PathBrowser.h"
 
@@ -134,6 +135,7 @@ protected:
 	void CreateViews() override;
 	void CreateRecentTab();
 	GameBrowser *CreateBrowserTab(const Path &path, std::string_view title, std::string_view howToTitle, std::string_view howToUri, BrowseFlags browseFlags, bool *bGridView, float *scrollPos);
+	UI::ViewGroup *CreateButtons(bool vertical);
 
 	void DrawBackground(UIContext &dc) override;
 	void update() override;
