+++ UI/TouchControlLayoutScreen.h
@@ -19,6 +19,7 @@
 
 #include "Common/UI/View.h"
 #include "Common/UI/ViewGroup.h"
+#include "Common/UI/TabHolder.h"
 #include "MiscScreens.h"
 
 class ControlLayoutView;
